Refactor static methods as functions in util module
